Completing specs and docs and tests
